requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
coverage==7.6.8
cycler==0.12.1
dateparser==1.2.0
distro==1.9.0
fonttools==4.55.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.7.1
kiwisolver==1.4.7
matplotlib==3.9.3
mplfinance==0.12.10b0
multidict==6.1.0
numpy==2.1.3
openai==0.28.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.0
pycryptodome==3.21.0
pydantic==2.10.0
pydantic_core==2.27.0
pyparsing==3.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.8.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
vulture==2.13
websockets==14.1
yarl==1.17.2
